+++ test/unit/fuzz-tools.test.js
@@ -122,6 +122,7 @@ import * as neuroV121 from '../../lib/neuro-v121.js';
 import * as neuroV122 from '../../lib/neuro-v122.js';
 import * as psychV123 from '../../lib/psych-v123.js';
 import * as hepV124 from '../../lib/hep-v124.js';
+import * as hepV125 from '../../lib/hep-v125.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -188,6 +189,7 @@ const MODULES = {
   'neuro-v122.js': neuroV122,
   'psych-v123.js': psychV123,
   'hep-v124.js': hepV124,
+  'hep-v125.js': hepV125,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];